more worker-like py interpreter packaging

+++ misc/packages/setuptools/wheel.py
@@ -1,7 +1,6 @@
 """Wheels support."""
 
 from distutils.util import get_platform
-from distutils import log
 import email
 import itertools
 import os
@@ -9,12 +8,11 @@ import posixpath
 import re
 import zipfile
 
-import pkg_resources
-import setuptools
-from pkg_resources import parse_version
-from setuptools.extern.packaging.tags import sys_tags
+from pkg_resources import Distribution, PathMetadata, parse_version
 from setuptools.extern.packaging.utils import canonicalize_name
 from setuptools.extern.six import PY3
+from setuptools import Distribution as SetuptoolsDistribution
+from setuptools import pep425tags
 from setuptools.command.egg_info import write_requirements
 
 
@@ -77,11 +75,11 @@ class Wheel:
 
     def is_compatible(self):
         '''Is the wheel is compatible with the current platform?'''
-        supported_tags = set((t.interpreter, t.abi, t.platform) for t in sys_tags())
+        supported_tags = pep425tags.get_supported()
         return next((True for t in self.tags() if t in supported_tags), False)
 
     def egg_name(self):
-        return pkg_resources.Distribution(
+        return Distribution(
             project_name=self.project_name, version=self.version,
             platform=(None if self.platform == 'any' else get_platform()),
         ).egg_name() + '.egg'
@@ -132,9 +130,9 @@ class Wheel:
         zf.extractall(destination_eggdir)
         # Convert metadata.
         dist_info = os.path.join(destination_eggdir, dist_info)
-        dist = pkg_resources.Distribution.from_location(
+        dist = Distribution.from_location(
             destination_eggdir, dist_info,
-            metadata=pkg_resources.PathMetadata(destination_eggdir, dist_info),
+            metadata=PathMetadata(destination_eggdir, dist_info),
         )
 
         # Note: Evaluate and strip markers now,
@@ -157,23 +155,17 @@ class Wheel:
             os.path.join(egg_info, 'METADATA'),
             os.path.join(egg_info, 'PKG-INFO'),
         )
-        setup_dist = setuptools.Distribution(
+        setup_dist = SetuptoolsDistribution(
             attrs=dict(
                 install_requires=install_requires,
                 extras_require=extras_require,
             ),
         )
-        # Temporarily disable info traces.
-        log_threshold = log._global_log.threshold
-        log.set_threshold(log.WARN)
-        try:
-            write_requirements(
-                setup_dist.get_command_obj('egg_info'),
-                None,
-                os.path.join(egg_info, 'requires.txt'),
-            )
-        finally:
-            log.set_threshold(log_threshold)
+        write_requirements(
+            setup_dist.get_command_obj('egg_info'),
+            None,
+            os.path.join(egg_info, 'requires.txt'),
+        )
 
     @staticmethod
     def _move_data_entries(destination_eggdir, dist_data):
@@ -213,8 +205,6 @@ class Wheel:
             for mod in namespace_packages:
                 mod_dir = os.path.join(destination_eggdir, *mod.split('.'))
                 mod_init = os.path.join(mod_dir, '__init__.py')
-                if not os.path.exists(mod_dir):
-                    os.mkdir(mod_dir)
-                if not os.path.exists(mod_init):
+                if os.path.exists(mod_dir) and not os.path.exists(mod_init):
                     with open(mod_init, 'w') as fp:
                         fp.write(NAMESPACE_PACKAGE_INIT)
